localSetup/projects/pubsub-init/requirements.txt
google-cloud-pubsub
Flask